fix adding documents
